+++ test/.wharf-ci.yml
@@ -6,6 +6,8 @@ test:
       image: alpine:latest
       cmds:
       - echo 'Hello world!'
+      - echo 'output from somefile.txt:'
+      - cat somefile.txt
   step2:
     container:
       image: busybox:latest
